move RevocationMatrixExplanation and change font-weight to 400
