+++ src/components/charts/new_revocations/RevocationMatrix/RevocationMatrixExplanation.js
@@ -17,13 +17,13 @@
 
 import React from "react";
 
-import { translate } from "../../../views/tenants/utils/i18nSettings";
+import { translate } from "../../../../views/tenants/utils/i18nSettings";
 
 const RevocationMatrixExplanation = () => (
   <div className="matrix-explanation bgc-white p-20">
     <h4>Using this chart</h4>
-    <p className="fw-300">{translate("matrixExplanationP1")}</p>
-    <p className="fw-300">{translate("matrixExplanationP2")}</p>
+    <p className="fw-400">{translate("matrixExplanationP1")}</p>
+    <p className="fw-400">{translate("matrixExplanationP2")}</p>
     <div className="d-f mT-20">
       <div className="example-icon-container">
         <div className="example-violation-total">35</div>
